PIA Network/backend/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.19
asyncpg==0.27.0
databases==0.6.3
python-telegram-bot==20.4
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==1.3.1
motor==4.2.0
httpx==0.24.1
google-auth==2.21.0
